New function called "WriteSolverInput" added to the Write module.
This functions allows the user to append the input parameters needed for the Solver module into the .msh file containing the Nodes and Elements of the PreProcessor Stage.

In this way, the solver module can be feeded with a file by reading the new data appended to the msh.

If the file has Solver input already appended at its end, this function errases the previous data and adds new data on top.