Solver module and instances defined. Up to now I added support for spectral solution of a equation of the form [K]{u] = lam[M]{u}.
Then I added an example file (QM_dirichlet_example.py) that shows how to run a simulation from the start to the writing of a vtk file.

test2.bc was added because I still have to implement bloch boundary conditions.

There is some doubt on how to build an interpreter that mixes both Dirichlet and Bloch boundaries.